Frontend work on the main, and map HTML files
I began changing the information in the layoutit basic layout, as far as the text is concerned.
I commented out a section of the navbar in the base, until we need it.
I added a map of the location on map.html, and the links to mapbox.
In urls, I gave the links names.